+++ pelutils/tests.py
@@ -10,6 +10,8 @@ from typing import Callable, TypeVar
 
 from pelutils import OS
 
+__all__ = ("SimplePool", "UnitTestCollection", "restore_argv")
+
 _C = TypeVar("_C", bound=Callable)  # pyright: ignore[reportMissingTypeArgument]